added pong directory
